remove base class from payload listener service
